// game/web/src/glossref.generated.ts
export interface MinionRef { name: string; id: string; hp: number | null; hpDynamic: boolean; elem: string; owner: string; }
// ...
export const MINION_REFS: MinionRef[] = [
  { name: "Prisma Azure", id: "trinityazureminion", hp: 65, hpDynamic: false, elem: "prism", owner: "Prisma Trinity" },
  { name: "Prisma Saffron", id: "trinitysaffronminion", hp: 65, hpDynamic: false, elem: "prism", owner: "Prisma Trinity" },
  { name: "Prisma Crimson", id: "trinitycrimsonminion", hp: 65, hpDynamic: false, elem: "prism", owner: "Prisma Trinity" },
  { name: "Angel", id: "ayanaangelminion", hp: 30, hpDynamic: false, elem: "angel", owner: "Ayana" },
  { name: "Grave", id: "blackknightgraveminion", hp: 30, hpDynamic: false, elem: "grave", owner: "Black Knight" },
  { name: "Skeleton", id: "blackknightskeletonminion", hp: 10, hpDynamic: false, elem: "lich", owner: "Black Knight" },
  { name: "Stonecap Mushroom", id: "rolandsporeminion", hp: 20, hpDynamic: false, elem: "spore", owner: "Roland" },
  { name: "Zombie", id: "blackknightzombieminion", hp: 40, hpDynamic: false, elem: "grave", owner: "Black Knight" },
  { name: "Seedling", id: "gaiaminion", hp: 25, hpDynamic: false, elem: "earth", owner: "Gaia" },
  { name: "Gaia, Enraged", id: "gaiamoonminion", hp: null, hpDynamic: true, elem: "moon", owner: "Gaia" },
  { name: "Mushroom", id: "gaiamushroomminion", hp: 20, hpDynamic: false, elem: "spore", owner: "Gaia" },
  { name: "World Tree", id: "gaiamythminion", hp: 80, hpDynamic: false, elem: "myth", owner: "Gaia" },
  { name: "Worldsprout", id: "gaiaworldsproutminion", hp: 40, hpDynamic: false, elem: "earth", owner: "Gaia" },
  { name: "Synthesizer", id: "hectorbatteryminion", hp: 40, hpDynamic: false, elem: "battery", owner: "Hector" },
  { name: "Dennis", id: "hectorminion", hp: 80, hpDynamic: false, elem: "poison", owner: "Hector" },
  { name: "Cinderling", id: "jarrikminion", hp: 10, hpDynamic: false, elem: "fire", owner: "Jarrik" },
  { name: "Azure Sparkling", id: "jarrikplasmaminion", hp: 10, hpDynamic: false, elem: "plasma", owner: "Jarrik" },
  { name: "Shadow Clone", id: "larianinjaminion", hp: 5, hpDynamic: false, elem: "ninja", owner: "Laria" },
  { name: "Bjorn, True King", id: "mythminion", hp: 80, hpDynamic: false, elem: "myth", owner: "" },
  { name: "Bjorn", id: "mythminion", hp: 80, hpDynamic: false, elem: "myth", owner: "" },
  { name: "Slimeball", id: "rolandslimeminion", hp: 30, hpDynamic: false, elem: "slime", owner: "Roland" },
  { name: "Slime", id: "riverdaughterslimeminion", hp: 15, hpDynamic: false, elem: "slime", owner: "River Daughter" },
  { name: "Boulder", id: "rolandminion", hp: 50, hpDynamic: false, elem: "earth", owner: "Roland" },
  { name: "Shady Assistant", id: "sayareanimationminion1", hp: 25, hpDynamic: false, elem: "reanimation", owner: "Saya" },
  { name: "Frozen Beast", id: "stasisminion", hp: 50, hpDynamic: false, elem: "stasis", owner: "" },
  { name: "Adult Eagle", id: "syladultminion", hp: 80, hpDynamic: false, elem: "wind", owner: "Syl" },
  { name: "Ancient Eagle", id: "sylancientminion", hp: 100, hpDynamic: false, elem: "wind", owner: "Syl" },
  { name: "Sparrowrider", id: "sylfaerieminion", hp: 15, hpDynamic: false, elem: "faerie", owner: "Syl" },
  { name: "Hatchling Eagle", id: "sylminion", hp: 60, hpDynamic: false, elem: "wind", owner: "Syl" },
  { name: "Scavenger's Pile", id: "sylnomadminion", hp: 25, hpDynamic: false, elem: "nomad", owner: "Syl" },
  { name: "Summer Courtesan", id: "titaniaminion", hp: 30, hpDynamic: false, elem: "poison", owner: "Titania" },
  { name: "Winter Loyalist", id: "titaniastasisminion", hp: 20, hpDynamic: false, elem: "stasis", owner: "Titania" },
];